+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * (theimo1221) Round battery value before db entry
 * (theimo1221) Allow zigbee device device query
+* (theimo1221) Trigger device query within EuroHeater from time to time automatically
 
 ## 1.2.0 (2022-10-22)